feat: noetbook to evaluate retrieval
